Remove BlocksKit

+++ thekiller/TKBluetoothManager.m
@@ -143,7 +143,7 @@ NSString* const CHARACTERISTIC_UUID = @"BD5DF558-9DF1-4216-8521-411D6F917A8C";
         [self willChangeValueForKey:@"nearbyDevicesDictionary"];
         [self didChangeValueForKey:@"nearbyDevicesDictionary"];
         
-        NSLog(@"Updated %@ with RSSI %d",device.name,device.rssi);
+        NSLog(@"Updated %@ with RSSI %ld",device.name,(long)device.rssi);
     }
 }
